microservice/requirements.txt

fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
openai==1.12.0
manim==0.19.0
markdown-it-py==3.0.0
pydantic==2.6.3
python-multipart==0.0.9
typing-extensions==4.10.0